Move some utility methods out of parameters.py
